Refactor
Routes are now a single file within an app.

Models are now a single file within an app.

Views have been renamed to templates.
