+++ LICENSE
@@ -1,6 +1,6 @@
 The MIT License
 
-Copyright (c) 2015 XMEN-Framework, http://xmen.io
+Copyright (c) 2016 XMEN-Framework, http://xmen.io
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
