+++ app/admin/templates/partials/footer.html
