+++ app/users/routes.js
@@ -5,5 +5,5 @@ var multipart = require('connect-multiparty');
 
 module.exports = function( app, passport, auth ) {
 
-    var users = require('../controllers/users');
+    var users = require('./controllers/users');
 };
